TrafficEdge has not to values for cars going from left to right node AND from right to left.
DecisionAlgorithm changed appropriately but needs to be verified.

Tried to add UI to display traffic graph - using simple JComponents of Swing - to display Nodes and Edges using JTables with numbers of cars waiting for manouevres and on the edges. Only Components coded - no graph display yet. This method can be changed however